+++ fixtures/text/sample.clj
@@ -1,0 +1,2 @@
+#!/usr/bin/clojure
+(println "Hello World")