+++ fixtures/text/sample.tcl
@@ -1,0 +1,2 @@
+#!/usr/bin/tclsh
+puts "Hello World"